requirements.txt
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.6.2.post1
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
binaryornot==0.4.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==44.0.0
daphne==4.1.2
distro==1.9.0
Django==5.1.3
django-browser-reload==1.17.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.16.1
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.159.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
jiter==0.8.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.56.1
packaging==24.2
pandas==2.2.3
pillow==11.3.0
promise==2.3
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
service-identity==24.2.0
setuptools==75.6.0
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
sympy==1.13.1
text-unidecode==1.3
tqdm==4.67.1
Twisted==24.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.35.0
whitenoise==6.9.0
zope.interface==7.2
